Update README.md
# Explorer

This is an open source project and an attempt to re-create the Windows Explorer's while exploring the capabilities of Vanilla JavaScript and attempting to stay as far away as possible from heavy frameworks.

This project was never meant to be finished, but rather an attempt to discover first-hand the problems of trying to write everything from scratch, and what it takes to make the development experience in JavaScript productive, reliable and enjoyable.

This project was built in a pair programming fashion where we discussed how we wanted to implement feature and how to improve every step of the way.

This project helped to discover why TypeScript is a necessity, the advantages of component based programming and reactive programming.

In a next iteration, we will be looking at implementing the project again from Scratch in React and Svelte. And attempting to cover all the remaining features in the project todo list.